requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
django==2.2.4
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-smart-selects==1.5.4
idna==2.8
iyzipay==1.0.32
oauthlib==3.1.0
pkg-resources==0.0.0
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
urllib3==1.25.3
sorl-thumbnail
Pillow

